+++ SqlDAL/Domain/Friend.cs
@@ -4,8 +4,8 @@ namespace SqlDAL.Domain
 {
     public class Friend
     {
-        public virtual int Id { get; set; }
-        public virtual int MemberId { get; set; }
+        public virtual long Id { get; set; }
+        public virtual long MemberId { get; set; }
 
         public virtual Member TFriend { get; set; }
